% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRIDGE_cov.R
\name{BRIDGE_cov}
\alias{BRIDGE_cov}
\title{BRIDGE_cov}
\usage{
BRIDGE_cov(
  infile,
  SubjectID,
  batch,
  time,
  mdl = NULL,
  data,
  imputation = FALSE,
  optimization = FALSE
)
}
\arguments{
\item{infile}{A genomic measure matrix (genes x observations) - for example, gene expression matrix.  \code{NA}s are not allowed}

\item{SubjectID}{A vector of the subject ID of the 'infile' data matrix. The samples are collected from the same subject having a unique ID.}

\item{batch}{A factor vector indicates each observation of the 'infile' data matrix are obtained from which batches. The labels of batch are \code{batch1} and \code{batch2}.}

\item{time}{A factor vector indicates each observation of the 'infile' data matrix are collected at which time points. The labels of time points are \code{time1} and \code{time2}.}

\item{mdl}{A A design matrix including covariates information.}

\item{data}{a data frame containing all information except expression data (observation x variables) including batch, time, SubjectID and other covariates.}

\item{imputation}{(Optical) Default value is \code{FALSE}. When \code{optimization = FALSE}, \code{imputation = FALSE} adjust data using empirical Bayes estimations of additive batch effects and multiplicative batch effects and after correcting the batch effects(Method 1), the adjusted data can be used for downstream analysis as if all samples are measured in batch 1; and \code{imputation = TRUE}, BRIDGE adjusts data based on linear regression imputation using empirical Bayes estimate of linear coefficients for the date of the non-bridging samples(Method 2). Method 2 works better in controlling type I error. After correcting the batch effects, the adjusted data can be used for downstream analysis as if all samples are measured in batch 2.}

\item{optimization}{(Optical) Default value is \code{FALSE}. If \code{TRUE}, BRIDGE automatically choose one method of \code{imputation} to adjusts data in order to achieve higher power (maybe) at the expense of inflated type I error.}
}
\value{
Function outputs a list including the following:
\describe{
\item{\code{factor.df}}{A data frame contains batch, time and subject id information}
\item{\code{AdjustedData}}{A genomic measure matrix (genes x observations) with batch, time and subjectID information at first 3 rows. If method 2 is used, then return data matrix adds new columns containing imputed data for non-bridging samples.}
\item{\code{phi2hat}}{A vector of J genes/features containing empirical Bayes estimate of additive batch effects og batch 2 for each feature, assuming phi1 = 0}
\item{\code{delta2hat}}{A vector of J genes/features containing empirical Bayes estimate of multiplicative batch effects og batch 2 for each feature, assuming delta1 = 1}
\item{\code{alphahat}}{A vector of J genes/features containing empirical Bayes estimate of linear coefficient}
\item{\code{betahat}}{A vector of J genes/features containing empirical Bayes estimate of linear coefficient}
}
}
\description{
Batch effect Reduction of mIcroarray data with Dependent samples usinG an Empirical Bayes approach (BRIDGE) - Covariates. To circumvent the limitations of existing batch-effect correction methods studies involving the longitudinal collection of high-dimensional 'omic' data, we propose Batch effect coRrectIon of microarray data with Dependent samples usinG an Empirical Bayes approach (BRIDGE). BRIDGE accounts for within-subject dependency expected in longitudinal studies with high-dimensional 'omic' data and involves the estimation and correction of additive batch and multiplicative batch effects when batch effects are confounded with time. After correcting for batch effects with BRIDGE, adjusted data can be used for downstream statistical analyses as if all samples are measured in a single batch. BRIDGE is applicable to different 'omics' platforms, such as microarray gene expression, DNA methylation or neuroimaging studies of neurodevelopment, as long as the transformed data are high-dimension and approximately normal distribution.
}
