% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwrBRIDGE_shiny.R
\name{pwrBRIDGE_shiny}
\alias{pwrBRIDGE_shiny}
\title{Shiny}
\usage{
pwrBRIDGE_shiny()
}
\value{
pwrBRIDGE_shiny user-interface
}
\description{
pwrBRIDGE pwrBRIDGE performs power analysis and bridging sample size calculation via empirical simulation for batch-confounded longitudinal study.
pwrBRIDGE generates data from multivariate normal distribution considering the subject and sample dependency with batch effects incorporating based
on location and scale (L/S) model. Then the batch effects are corrected using any of BRIDGE, ComBat and longitudinal ComBat. The corrected data are
used for downstream hypothesis testing and power/bridging sample size assessment.
}
